frontend/requirements.txt
streamlit
# FastAPI backend
fastapi==0.113.0
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.22
psycopg2-binary==2.9.11  # precompiled binary, avoids C++ build issues

# Environment variables
python-dotenv==1.0.1

# Pydantic for validation
pydantic==2.12.4  # compatible with Python 3.12
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Security
bcrypt==4.1.1

# Testing
pytest==8.4.0
httpx==0.28.1  # for API testing

# Other
requests==2.32.4
sentry-sdk==2.8.0
zipp==3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=1.26.19 # not directly required, pinned by Snyk to avoid a vulnerability

sqlalchemy-bigquery
google-cloud-bigquery
email-validator
alembic